Bugfix concerning hangs on game start, APAX added
Lobbyhelper hangs the server while it's joining the background
profile lookup thread. This would previously be done at every game start.

This is usually not noticeable when using a fully populated database,
but if the database has been cleaned this often hangs the server
until the current HTTP request times out (20 seconds) due to the fact
that TF2Lobby profile lookups are slow and often time out.

This has been fixed by no longer stopping and joining the lookup thread
when the game starts.

Also, APAX (Asynchronous Pawn And XML) has been added. This is a simple
equivalent for the XMLHttpRequest object used for AJAX in web pages.
It can be used to easily do web requests and retrieve the results
asynchronously using callbacks.

Note that you must set a timer yourself to periodically call the
APAX_CheckResponses function.

When using APAX, make sure your script memory space is large enough to
hold the response. This can be done by compiling with ./spcomp -S<memsize>